@@ -8,11 +8,12 @@
     "@fortawesome/free-solid-svg-icons": "5.15.1",
     "@fortawesome/react-fontawesome": "0.1.13",
     "classnames": "2.2.6",
+    "eslint-plugin-simple-import-sort": "^7.0.0",
     "firebase": "9.0.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-router-dom": "6.1.1",
-    "react-scripts": "4.0.1",
+    "react-scripts": "5.0.1",
     "sass": "1.48.0",
     "web-vitals": "1.0.1"
   },
@@ -31,20 +32,17 @@
     "@types/firebase": "3.2.1",
     "@types/jest": "26.0.15",
     "@types/node": "12.0.0",
-    "@types/react": "17.0.0",
-    "@types/react-dom": "17.0.0",
-    "@types/react-router-dom": "5.3.2",
-    "babel-eslint": "10.1.0",
-    "eslint-config-airbnb-typescript-prettier": "3.1.0",
-    "eslint-import-resolver-typescript": "2.3.0",
-    "eslint-plugin-react": "7.21.5",
-    "eslint-plugin-react-hooks": "4.2.0",
+    "@types/react-dom": "^18.0.5",
+    "@typescript-eslint/eslint-plugin": "^5.25.0",
+    "@typescript-eslint/parser": "^5.27.0",
+    "eslint": "^8.15.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-import-resolver-typescript": "^2.7.1",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-react": "^7.29.4",
     "husky": "4.2.5",
-    "prettier": "2.0.5",
-    "typescript": "4.0.3"
-  },
-  "resolutions": {
-    "ini": "1.3.7"
+    "prettier": "^2.6.2",
+    "typescript": "^4.7.3"
   },
   "husky": {
     "hooks": {
@@ -61,7 +59,6 @@
     "storybook": "start-storybook -p 6006 -s public",
     "build-storybook": "build-storybook -s public",
     "deploy": "yarn build && firebase deploy --only hosting"
-
   },
   "eslintConfig": {
     "extends": [
