@@ -1,5 +1,6 @@
 import React from 'react';
-import { ComponentStory, ComponentMeta } from '@storybook/react';
+import { ComponentStory } from '@storybook/react';
+
 import Button from 'components/button/Button';
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
@@ -9,7 +10,7 @@ export default {
 };
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof Button> = args => <Button>Submit</Button>;
+const Template: ComponentStory<typeof Button> = () => <Button>Submit</Button>;
 
 export const Primary = Template.bind({});
 // // More on args: https://storybook.js.org/docs/react/writing-stories/args
