@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
-import classnames from 'classnames';
 import { Link } from 'react-router-dom';
+import classnames from 'classnames';
 
 import Anchor from 'components/anchor/Anchor';
 
@@ -16,11 +16,15 @@ type Props = {
 const Header: FC<Props> = ({ route = '/', href, title, className }) => (
   <div className={classnames(styles.title, className)}>
     {!href ? (
-      <Link className={styles.header} to={route}>
+      <Link
+        className={styles.header}
+        to={route}>
         {title}
       </Link>
     ) : (
-      <Anchor className={styles.header} href={href}>
+      <Anchor
+        className={styles.header}
+        href={href}>
         {title}
       </Anchor>
     )}
