@@ -1,17 +1,13 @@
-import React, { FC, useRef, useState } from 'react';
-import classnames from 'classnames';
+import React, { FC } from 'react';
+import { faGithubAlt, faLinkedinIn } from '@fortawesome/free-brands-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import {
-  faGithubAlt,
-  faGitlab,
-  faLinkedinIn,
-} from '@fortawesome/free-brands-svg-icons';
-import { faAt } from '@fortawesome/free-solid-svg-icons';
+import classnames from 'classnames';
 
 import Anchor from 'components/anchor/Anchor';
 
-import styles from './Header.module.scss';
 import Nav from './nav/Nav';
+
+import styles from './Header.module.scss';
 
 const Header: FC = () => {
   return (
@@ -21,13 +17,14 @@ const Header: FC = () => {
         <Nav />
       </div>
       <div className={styles.socials}>
-        <Anchor href="https://www.github.com/jeanaica" className={styles.link}>
+        <Anchor
+          href='https://www.github.com/jeanaica'
+          className={styles.link}>
           <FontAwesomeIcon icon={faGithubAlt} />
         </Anchor>
         <Anchor
-          href="https://www.linkedin.com/in/jeanaicasuplido/"
-          className={styles.link}
-        >
+          href='https://www.linkedin.com/in/jeanaicasuplido/'
+          className={styles.link}>
           <FontAwesomeIcon icon={faLinkedinIn} />
         </Anchor>
       </div>
