@@ -1,5 +1,5 @@
-import React, { Dispatch, RefObject, SetStateAction } from 'react';
-import { Link, useLocation } from 'react-router-dom';
+import React, { FC, RefObject } from 'react';
+import { Link } from 'react-router-dom';
 
 import styles from './Dropdown.module.scss';
 
@@ -8,24 +8,24 @@ type Props = {
   onClick(): void;
 };
 
-const Dropdown: React.FC<Props> = ({ forwardRef, onClick }) => (
-  <ul className={styles['nav-dropdown']} ref={forwardRef}>
+const Dropdown: FC<Props> = ({ forwardRef, onClick }) => (
+  <ul
+    className={styles['nav-dropdown']}
+    ref={forwardRef}>
     <li>
-      <Link to="/about" onClick={onClick}>
+      <Link
+        to='/about'
+        onClick={onClick}>
         About
       </Link>
     </li>
-    {/* <li>
-        <Link to="/blog">Blog</Link>
-      </li> */}
     <li>
-      <Link to="/portfolio" onClick={onClick}>
+      <Link
+        to='/portfolio'
+        onClick={onClick}>
         Portfolio
       </Link>
     </li>
-    {/* <li>
-      <Link to="/contact">Contact</Link>
-    </li> */}
   </ul>
 );
 
