@@ -1,11 +1,11 @@
-import React, { FC, ReactElement, HTMLAttributes } from 'react';
+import React, { FC, HTMLAttributes } from 'react';
 import classnames from 'classnames';
 
 import styles from './SectionTitle.module.scss';
 
 type Props = {
   text: string;
-  highlight: string;
+  highlight?: string;
   className?: string;
 } & HTMLAttributes<HTMLDivElement>;
 
@@ -15,7 +15,9 @@ const SectionTitle: FC<Props> = ({
   className,
   ...rest
 }: Props) => (
-  <div className={classnames(className, styles.section)} {...rest}>
+  <div
+    className={classnames(className, styles.section)}
+    {...rest}>
     <h5 className={styles.title}>{text}</h5>
     <div className={styles.accent}>
       <h5 className={styles.highlight}>{highlight}</h5>
