@@ -1,10 +1,10 @@
 import React, { FC } from 'react';
 
 import Card from 'components/card/Card';
-import Header from 'components/card/header/Header';
 import Content from 'components/card/content/Content';
-import Summary from 'components/card/summary/Summary';
+import Header from 'components/card/header/Header';
 import Preview from 'components/card/preview/Preview';
+import Summary from 'components/card/summary/Summary';
 
 type Props = {
   title: string;
@@ -17,14 +17,21 @@ type Props = {
 
 const ProjectCard: FC<Props> = ({ title, text, link, href, img, right }) => (
   <Card right={right}>
-    <Preview right={right} img={img} href={href} />
+    <Preview
+      right={right}
+      img={img}
+      href={href}
+    />
     <Content right={right}>
-      <Header title={title} href={href} />
+      <Header
+        title={title}
+        href={href}
+      />
       <Summary
         right={right}
         text={text}
         link={link}
-        linkText="Go to repository >>"
+        linkText='Go to repository >>'
       />
     </Content>
   </Card>
